feat: added option to disable unit conversions for expressions that are not dimensionally correct
Sometimes we have expressions that are not dimensionally correct. The
input values are assumed to be expressed in a certain unit, and the
result is then the value expressed in a certain unit.

Added a no_unit_conversion option to the Expression.eval(...) function
to disable unit conversion so we can support these types of expressions.